Fix clippy::unnecessary_cast on Windows (WPARAM.0 is already usize)
Surfaced by CI's windows-latest runner -- the real build itself compiled
clean on the first try; only this cosmetic lint failed.

+++ crates/observation/src/windows/shortcuts.rs
@@ -137,7 +137,7 @@ impl GlobalShortcutSource {
                 break;
             }
             if msg.message == WM_HOTKEY {
-                let index = msg.wParam.0 as usize;
+                let index = msg.wParam.0;
                 if let Some(binding) = bindings.get(index) {
                     let _ = fired_tx.send(binding.label.clone());
                 }
